LinearSOS restructure
