+++ src/core/linsos/__init__.py
@@ -1,19 +1,13 @@
 from .linsos import LinearSOS
 from .tangents import root_tangents
 from .basis import (
-    LinearBasis, LinearBasisCyclic, LinearBasisTangentCyclic, LinearBasisAMGMCyclic,
-    CachedCommonLinearBasisSpecialCyclic, CachedCommonLinearBasisTangentCyclic,
-    LinearBasisTangent,
-    CachedCommonLinearBasisTangent
+    LinearBasis, LinearBasisTangent
 )
 from .solution import SolutionLinear
 
 __all__ = [
     'LinearSOS',
     'root_tangents',
-    'LinearBasis', 'LinearBasisCyclic', 'LinearBasisTangentCyclic', 'LinearBasisAMGMCyclic',
-    'CachedCommonLinearBasisSpecialCyclic', 'CachedCommonLinearBasisTangentCyclic',
-    'LinearBasisTangent',
-    'CachedCommonLinearBasisTangent',
+    'LinearBasis', 'LinearBasisTangent',
     'SolutionLinear'
 ]
