+++ src/core/linsos/updegree.py
@@ -2,10 +2,9 @@ from typing import Generator, Dict
 
 import sympy as sp
 from sympy.core.singleton import S
+from sympy.combinatorics import PermutationGroup
 
-from .basis import LinearBasis, LinearBasisCyclic, a, b, c
-from ...utils.polytools import deg
-from ...utils.expression.cyclic import CyclicSum, CyclicProduct
+from .basis import LinearBasis
 
 
 class LinearBasisMultiplier(LinearBasis):
@@ -15,34 +14,22 @@ class LinearBasisMultiplier(LinearBasis):
     then it is equivalent to
     \sum (a^2 - ab) * f(a,b,c) = RHS + x0 * \sum -ab * f(a,b,c).
 
-    This converts the problem to an usual linear programming by adding a basis \sum -ab * f(a,b,c).
+    This converts the problem to a usual linear programming by adding a basis \sum -ab * f(a,b,c).
     """
-    is_cyc = False
     def __init__(self, poly, multiplier):
         self.poly = poly
         self.multiplier = multiplier
-        self.expr_ = poly * (-multiplier.doit().as_poly(a,b,c))
-        self.array_ = None
-        self.array_sp_ = None
-
-class LinearBasisMultiplierCyclic(LinearBasisCyclic, LinearBasisMultiplier):
-    r"""
-    For example, if we want to find
-    \sum (a^2 + x0*ab) * f(a,b,c) = \sum (x1 * g1(a,b,c) + x2 * g2(a,b,c) + ...)
-    then it is equivalent to
-    \sum (a^2 - ab) * f(a,b,c) = RHS + x0 * \sum -ab * f(a,b,c).
-
-    This converts the problem to an usual linear programming by adding a basis \sum -ab * f(a,b,c).
-    """
-    is_cyc = True
-    def __init__(self, *args, **kwargs):
-        super(LinearBasisMultiplierCyclic, self).__init__(*args, **kwargs)
+    def as_poly(self, symbols) -> sp.Poly:
+        poly = (self.poly * (-self.multiplier.doit().as_poly(self.poly.gens)))
+        return poly.xreplace(dict(zip(self.poly.gens, symbols)))
+    def as_expr(self, symbols) -> sp.Expr:
+        return self.as_poly(symbols).as_expr()
 
 
-def higher_degree(
-        poly: sp.polys.Poly,
+def lift_degree(
+        poly: sp.Poly,
         degree_limit: int = 12,
-        is_cyc: bool = True
+        symmetry: PermutationGroup = PermutationGroup(),
     ) -> Generator[Dict, None, None]:
     """
     Hilbert's problem has shown that not every positive polynomial can be written as a sum of squares.
@@ -51,7 +38,7 @@ def higher_degree(
     f(a,b,c) * h(a,b,c) = g(a,b,c).
 
     In practice, we can try out h(a,b,c) = \sum a, h(a,b,c) = \sum (a^2-ab + xab) and so on.
-    This `higher_degree` function would generate the h(a,b,c) and associated information.
+    This `lift_degree` function would generate the h(a,b,c) and associated information.
 
     Parameters
     ----------
@@ -78,35 +65,37 @@ def higher_degree(
         add_degree: int
             The degree of h(a,b,c).
     """
-    n = deg(poly)
+    n = poly.total_degree()
     n_plus = 0
 
-    if is_cyc:
-        multipliers = {
-            1: CyclicSum(a),
-            2: CyclicSum(a**2 - a*b),
-            3: CyclicSum(a*(a-b)*(a-c)),
-            4: CyclicSum(a**2*(a-b)*(a-c)),
-        }
+    # if is_cyc:
+    #     multipliers = {
+    #         1: CyclicSum(a),
+    #         2: CyclicSum(a**2 - a*b),
+    #         3: CyclicSum(a*(a-b)*(a-c)),
+    #         4: CyclicSum(a**2*(a-b)*(a-c)),
+    #     }
 
-        adjustment_multipliers = {
-            1: [],
-            2: [CyclicSum(a*b)],
-            3: [CyclicSum(a**2*b), CyclicSum(a*b**2), CyclicSum(a*b*c)],
-            4: [CyclicSum(a**3*b), CyclicSum(a*b**3), CyclicSum(a**2*b*c), CyclicSum(a*b*(a-b)**2)],
-        }
-    else:
-        multipliers = {
-            1: c,
-            2: (a-b)**2
-        }
+    #     adjustment_multipliers = {
+    #         1: [],
+    #         2: [CyclicSum(a*b)],
+    #         3: [CyclicSum(a**2*b), CyclicSum(a*b**2), CyclicSum(a*b*c)],
+    #         4: [CyclicSum(a**3*b), CyclicSum(a*b**3), CyclicSum(a**2*b*c), CyclicSum(a*b*(a-b)**2)],
+    #     }
+    # else:
+    #     multipliers = {
+    #         1: c,
+    #         2: (a-b)**2
+    #     }
 
-        adjustment_multipliers = {
-            1: [a, b],
-            2: [(b-c)**2, (a-c)**2, a*b, a*c, b*c]
-        }
+    #     adjustment_multipliers = {
+    #         1: [a, b],
+    #         2: [(b-c)**2, (a-c)**2, a*b, a*c, b*c]
+    #     }
+    multipliers = {}
+    adjustment_multipliers = {}
 
-    while n + n_plus <= degree_limit and n_plus <= max(multipliers.keys()):
+    while n + n_plus <= degree_limit and n_plus <= max([0] + list(multipliers.keys())):
         if n_plus == 0:
             yield {
                 'poly': poly,
@@ -118,11 +107,9 @@ def higher_degree(
             n_plus += 1
             continue
 
-        basis_type = LinearBasisMultiplierCyclic if is_cyc else LinearBasisMultiplier
-
-        new_poly = poly * multipliers[n_plus].doit().as_poly(a,b,c)
+        new_poly = poly * multipliers[n_plus].doit().as_poly(poly.gens)
         adjustment_basis = [
-            basis_type(poly, multiplier) for multiplier in adjustment_multipliers[n_plus]
+            LinearBasisMultiplier(poly, multiplier) for multiplier in adjustment_multipliers[n_plus]
         ]
 
         yield {
