+++ src/core/structsos/ternary/cubic.py
@@ -624,7 +624,7 @@ def _sos_struct_nonhom_cubic_symmetric(coeff, real = False):
 
         x = rationalize_func(sp.Poly([9, -12*c2, 4*c100], sp.Symbol('x')), validation=check_x,
                             validation_initial=lambda x: x >= 2*c2/3, direction = 1)
-    print(x, get_const(x), 't=',radsimp(c100 / (3*x**2/4)))
+    # print(x, get_const(x), 't =', radsimp(c100 / (3*x**2/4)))
     if x is None:
         return None
     res = get_const(x)
