+++ src/utils/expression/__init__.py
@@ -1,4 +1,4 @@
-from .coeff import Coeff
+from .coeff import Coeff, identify_symmetry
 
 from .cyclic import (
     CyclicExpr, CyclicSum, CyclicProduct, SymmetricSum, SymmetricProduct,
@@ -15,7 +15,7 @@ from .solution import (
 )
 
 __all__ = [
-    'Coeff',
+    'Coeff', 'identify_symmetry',
     'CyclicExpr', 'CyclicSum', 'CyclicProduct', 'SymmetricSum', 'SymmetricProduct', 'is_cyclic_expr',
     'poly_get_factor_form', 'poly_get_standard_form', 'latex_coeffs',
     'Solution', 'SolutionSimple',
